refactor: 'https://api.stlouisfed.org/fred/' baseUrl 변수 처리
